pyproject.toml
[project]
name = "dog-emotion-swin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "evaluate>=0.4.3",
    "grad-cam>=1.5.5",
    "matplotlib>=3.10.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "scikit-learn>=1.7.0",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "transformers>=4.52.4",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
